+++ components/css.ts
@@ -1,0 +1,18 @@
+import {StyleSheet} from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    alignItems: 'center',
+    justifyContent: 'center',
+    marginTop: 100,
+  },
+  appTitle: {
+    fontSize: 30,
+    fontWeight: 'bold',
+  },
+  input: {
+    width: 100,
+  },
+});
+
+export default styles;
